[ fix ] 입력하는 변수들 타입 병합
