+++ api/wishes/wishesAPI.ts
@@ -1,9 +1,8 @@
-import { AccountInfoType } from '@/types/wishes/accountInfotype';
 import PATH from '../../constant/path';
 import { client } from '../common/axios';
-import { WishesDataType } from '@/types/wishes/wishesDataType';
 import axios from 'axios';
 import { SiteDataType } from '@/types/siteDataType';
+import { AccountInfoType, WishesDataType } from '@/types/common/input';
 
 export const createWishesLink = async (wishesData: WishesDataType) => {
   const accessToken = localStorage.getItem('accessToken');
