--- types/wishes/accountInfotype.ts
@@ -1,5 +1,0 @@
-export interface AccountInfoType {
-  name: string;
-  bankName: string;
-  account: string;
-}
