--- types/wishes/wishesDataType.ts
@@ -1,10 +1,0 @@
-export interface WishesDataType {
-  imageURL: string;
-  price: number;
-  title: string;
-  hint: string;
-  initial: string;
-  startDate: Date;
-  endDate: Date;
-  phone: string;
-}
